split methods for scraping function
